Some behaviour modifications
Modified the air to turn into water near the map edge, allowing for the ocean to
'flood' in for an eternal supply of water. Also modified the gravity behaviour
to apply to stone and dirt, preventing blocks from hovering in mid-air. It sorta
works, and needs more effort to be complete.