Use antlr to parse the permissions file
This introduces a hard dependency on antlr3, but at the same time cleans up the
Policy code a ton. build.xml was updated to include the generation of the antlr
parser and lexer, along with modifying how Constants.java is produced with the
version info. Generated files now show up in gen/ and are added to the
compilation, keeping src/ clean of mechanical cruft.